+++ src/Reader/Collection.php
@@ -3,7 +3,7 @@
  * Zend Framework (http://framework.zend.com/)
  *
  * @link      http://github.com/zendframework/zf2 for the canonical source repository
- * @copyright Copyright (c) 2005-2015 Zend Technologies USA Inc. (http://www.zend.com)
+ * @copyright Copyright (c) 2005-2019 Zend Technologies USA Inc. (http://www.zend.com)
  * @license   http://framework.zend.com/license/new-bsd New BSD License
  */
 
@@ -11,6 +11,11 @@ namespace Zend\Feed\Reader;
 
 use ArrayObject;
 
+/**
+ * @deprecated This class is deprecated. Use the concrete collection classes
+ *     \Zend\Feed\Reader\Collection\Author and \Zend\Feed\Reader\Collection\Category
+ *     or the generic class \Zend\Feed\Reader\Collection\Collection instead.
+ */
 class Collection extends ArrayObject
 {
 }